+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.0]
+## [0.4.0] - 2019-05-22
+
+### Added
+- Auto-initialization of Heap for both iOS and Android apps.
+- Added `HeapIgnore` and `HeapIgnoreTargetText` components.
+
+### Fixed
+- Prevent Heap library errors from crashing the client app.
+- Passthrough props on React Navigation autocapture HOC.
+
+## [0.3.0] - 2019-04-23
 
 ### Added
 - Autotrack for React Native and NativeBase `Switch` components.
@@ -16,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Get Android+Detox tests working after breaking changes in a prior Detox upgrade.
 
-## [0.2.0]
+## [0.2.0] - 2019-03-29
 
 ### Added